+++ src/vmm/src/devices/virtio/block/device.rs
@@ -8,8 +8,9 @@ use super::BlockError;
 use super::persist::{BlockConstructorArgs, BlockState};
 use super::vhost_user::device::{VhostUserBlock, VhostUserBlockConfig};
 use super::virtio::device::{VirtioBlock, VirtioBlockConfig};
-use crate::devices::virtio::device::{IrqTrigger, VirtioDevice};
+use crate::devices::virtio::device::VirtioDevice;
 use crate::devices::virtio::queue::Queue;
+use crate::devices::virtio::transport::mmio::IrqTrigger;
 use crate::devices::virtio::{ActivateError, TYPE_BLOCK};
 use crate::rate_limiter::BucketUpdate;
 use crate::snapshot::Persist;
