+++ src/vmm/src/devices/virtio/mod.rs
@@ -18,12 +18,12 @@ pub mod device;
 pub mod generated;
 mod iov_deque;
 pub mod iovec;
-pub mod mmio;
 pub mod net;
 pub mod persist;
 pub mod queue;
 pub mod rng;
 pub mod test_utils;
+pub mod transport;
 pub mod vhost_user;
 pub mod vhost_user_metrics;
 pub mod vsock;
