+++ src/vmm/src/devices/virtio/net/test_utils.rs
@@ -104,7 +104,7 @@ impl TapTrafficSimulator {
 
         let send_addr_ptr = &mut storage as *mut libc::sockaddr_storage;
 
-        // SAFETY: `sock_addr` is a valid pointer and safe to derference.
+        // SAFETY: `sock_addr` is a valid pointer and safe to dereference.
         unsafe {
             let sock_addr: *mut libc::sockaddr_ll = send_addr_ptr.cast::<libc::sockaddr_ll>();
             (*sock_addr).sll_family = libc::sa_family_t::try_from(libc::AF_PACKET).unwrap();
@@ -223,7 +223,7 @@ pub fn if_index(tap: &Tap) -> i32 {
 
 /// Enable the tap interface.
 pub fn enable(tap: &Tap) {
-    // Disable IPv6 router advertisment requests
+    // Disable IPv6 router advertisement requests
     Command::new("sh")
         .arg("-c")
         .arg(format!(
@@ -311,7 +311,7 @@ pub mod test {
     use event_manager::{EventManager, SubscriberId, SubscriberOps};
 
     use crate::check_metric_after_block;
-    use crate::devices::virtio::device::{IrqType, VirtioDevice};
+    use crate::devices::virtio::device::VirtioDevice;
     use crate::devices::virtio::net::device::vnet_hdr_len;
     use crate::devices::virtio::net::generated::ETH_HLEN;
     use crate::devices::virtio::net::test_utils::{
@@ -320,6 +320,7 @@ pub mod test {
     use crate::devices::virtio::net::{MAX_BUFFER_SIZE, Net, RX_INDEX, TX_INDEX};
     use crate::devices::virtio::queue::{VIRTQ_DESC_F_NEXT, VIRTQ_DESC_F_WRITE};
     use crate::devices::virtio::test_utils::{VirtQueue, VirtqDesc};
+    use crate::devices::virtio::transport::mmio::IrqType;
     use crate::logger::IncMetric;
     use crate::vstate::memory::{Address, Bytes, GuestAddress, GuestMemoryMmap};
