+++ src/vmm/src/devices/virtio/persist.rs
@@ -12,8 +12,8 @@ use serde::{Deserialize, Serialize};
 use super::queue::QueueError;
 use crate::devices::virtio::device::VirtioDevice;
 use crate::devices::virtio::generated::virtio_ring::VIRTIO_RING_F_EVENT_IDX;
-use crate::devices::virtio::mmio::MmioTransport;
 use crate::devices::virtio::queue::Queue;
+use crate::devices::virtio::transport::mmio::MmioTransport;
 use crate::snapshot::Persist;
 use crate::vstate::memory::{GuestAddress, GuestMemoryMmap};
 
@@ -261,10 +261,10 @@ mod tests {
     use crate::devices::virtio::block::virtio::VirtioBlock;
     use crate::devices::virtio::block::virtio::device::FileEngineType;
     use crate::devices::virtio::block::virtio::test_utils::default_block_with_path;
-    use crate::devices::virtio::mmio::tests::DummyDevice;
     use crate::devices::virtio::net::Net;
     use crate::devices::virtio::net::test_utils::default_net;
     use crate::devices::virtio::test_utils::default_mem;
+    use crate::devices::virtio::transport::mmio::tests::DummyDevice;
     use crate::devices::virtio::vsock::{Vsock, VsockUnixBackend};
     use crate::snapshot::Snapshot;
