+++ src/vmm/src/devices/virtio/transport/mod.rs
@@ -1,0 +1,5 @@
+// Copyright 2025 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+/// MMIO transport for VirtIO devices
+pub mod mmio;
